just added the bee. going to apply scripts later on
hoping to add in a brand new scene eventually with a map. just focusing on scripts for the enemies first however. if a bee can sting you on a completely flat terrain, then adjusting what's necessary to have the bee follow you on more complicated terrain shouldn't be as difficult realistically.